Fixed a broken test, added basket from product view - stopping point for other work
