+++ payments/urls.py
@@ -1,9 +1,5 @@
 """URLs for the payments app."""
 
-from django.urls import include, re_path
+from payments.views.v0.urls import urlpatterns as v0_urls
 
-from payments.views.v0.urls import router as v0_router
-
-urlpatterns = [
-    re_path(r"^v0/", include(v0_router.urls)),
-]
+urlpatterns = v0_urls
