+++ payments/views/v0/urls.py
@@ -1,5 +1,7 @@
 """Routes specific to this version of the payments API."""
-from payments.views.v0.views import BasketItemViewSet, BasketViewSet
+from django.urls import include, re_path
+
+from payments.views.v0 import BasketItemViewSet, BasketViewSet, create_from_product
 from unified_ecommerce.routers import SimpleRouterWithNesting
 
 router = SimpleRouterWithNesting()
@@ -12,3 +14,12 @@ basket_router.register(
     basename="basket-items",
     parents_query_lookups=["basket"],
 )
+
+urlpatterns = [
+    re_path(
+        r"baskets/create_from_product/<str:system_slug>/<str:sku>/",
+        create_from_product,
+        name="create_from_product",
+    ),
+    re_path("^", include(router.urls)),
+]
